requirements.txt
torch-directml
git+https://github.com/trupljan/lightning
captcha
pyyaml
pillow
tqdm
tensorboard
tensorboardx
wheel>=0.38.0 # not directly required, pinned by Snyk to avoid a vulnerability